Fix description of inheritance hierarchies
The description was wrong in that it had `StructType` both as a meta type for creating concrete struct types, and as the base type for struct type instances.

In #22, @phpnode started a discussion about this topic, and while in that issue I argued for another solution - making `StructType` the base class for struct type instances, I've now come around to the point of view that having distinct hierarchies for meta-classes (not only `StructType`, but also `ValueType` once that becomes a thing), and type definitions and their instances makes sense. We should be able to talk about both types and their instances as objects, and those objects should have types.

I also introduced `%Type%`, `%Struct%`, and `%Struct%.array` as base-classes for those hierarchies. These are modeled on the precedence of `%TypedArray%`, but it might make sense to just expose the as `Type`, `Struct`, and (obviously) `%Struct%.array`.

I think it makes sense to go over the explainer and change all `FooType` constructor names to just `Foo` - the fact that they're types isn't all that special at this level: they're just constructors just as `function Point(){}` would be a constructor.